+++ x-pack/platform/test/fleet_api_integration/apis/epm/index.ts
@@ -56,5 +56,6 @@ export default function ({ loadTestFile, getService }: FtrProviderContext) {
     loadTestFile(require.resolve('./data_views'));
     loadTestFile(require.resolve('./custom_integrations'));
     loadTestFile(require.resolve('./rollback'));
+    loadTestFile(require.resolve('./knowledge_base'));
   });
 }
