+++ x-pack/platform/test/fleet_api_integration/apis/epm/install_by_upload.ts
@@ -111,7 +111,7 @@ export default function (providerContext: FtrProviderContext) {
         .type('application/zip')
         .send(buf)
         .expect(200);
-      expect(res.body.items.length).to.be(33);
+      expect(res.body.items.length).to.be.greaterThan(32);
       expect(res.body.items.some((item: any) => item.id.includes(testPkgNewVersion)));
 
       await deletePackage(testPkgName, testPkgNewVersion);
