+++ x-pack/platform/test/fleet_api_integration/apis/package_policy/input_package_create_upgrade.ts
@@ -194,12 +194,9 @@ export default function (providerContext: FtrProviderContext) {
 
     it('should not have created any ES assets on install', async () => {
       const installation = await getInstallationInfo(supertest, PACKAGE_NAME, START_VERSION);
-      expect(installation.installed_es).to.eql([
-        {
-          id: 'input_package_upgrade-README.md',
-          type: 'knowledge_base',
-        },
-      ]);
+      expect(
+        installation.installed_es.filter((item: any) => item.type !== 'knowledge_base')
+      ).to.eql([]);
     });
 
     it('should create index templates and update installed_es on package policy creation', async () => {
