+++ x-pack/platform/test/fleet_api_integration/apis/package_policy/upgrade.ts
@@ -9,14 +9,9 @@ import type {
   UpgradePackagePolicyDryRunResponse,
   UpgradePackagePolicyResponse,
 } from '@kbn/fleet-plugin/common/types';
-import { sortBy } from 'lodash';
 import type { FtrProviderContext } from '../../../api_integration/ftr_provider_context';
 import { skipIfNoDockerRegistry } from '../../helpers';
 import { getInstallationInfo } from './helper';
-
-const expectIdArraysEqual = (arr1: any[], arr2: any[]) => {
-  expect(sortBy(arr1, 'id')).to.eql(sortBy(arr2, 'id'));
-};
 
 export default function (providerContext: FtrProviderContext) {
   const { getService } = providerContext;
@@ -25,6 +20,7 @@ export default function (providerContext: FtrProviderContext) {
   const kibanaServer = getService('kibanaServer');
   const es = getService('es');
   const fleetAndAgents = getService('fleetAndAgents');
+  const retry = getService('retry');
 
   function withTestPackage(name: string, version: string) {
     const pkgRoute = `/api/fleet/epm/packages/${name}/${version}`;
@@ -51,8 +47,7 @@ export default function (providerContext: FtrProviderContext) {
     }
   };
 
-  // Failing: See https://github.com/elastic/kibana/issues/246204
-  describe.skip('Package Policy - upgrade', function () {
+  describe('Package Policy - upgrade', function () {
     skipIfNoDockerRegistry(providerContext);
     let agentPolicyId: string;
     let packagePolicyId: string;
@@ -1250,7 +1245,6 @@ export default function (providerContext: FtrProviderContext) {
         }
 
         expectedAssets.push({ id: 'logs@custom', type: 'component_template' });
-        expectedAssets.push({ id: 'integration_to_input-README.md', type: 'knowledge_base' });
         expectedAssets.push({ id: 'integration_to_input@custom', type: 'component_template' });
       });
 
@@ -1285,12 +1279,20 @@ export default function (providerContext: FtrProviderContext) {
             })
             .expect(200);
 
-          const installation = await getInstallationInfo(
-            supertest,
-            'integration_to_input',
-            '3.0.0'
-          );
-          expectIdArraysEqual(installation.installed_es, expectedAssets);
+          await retry.tryForTime(10000, async () => {
+            const installation = await getInstallationInfo(
+              supertest,
+              'integration_to_input',
+              '3.0.0'
+            );
+            expectedAssets.forEach((item) => {
+              expect(
+                installation.installed_es.find(
+                  (asset: any) => asset.type === item.type && asset.id === item.id
+                )
+              ).to.eql(item, `Expected asset not found: ${item.type}:${item.id}`);
+            });
+          });
 
           const expectedComponentTemplates = expectedAssets.filter(
             (expectedAsset) =>
